todo_backend/requirements.txt
certifi==2018.10.15
chardet==3.0.4
defusedxml==0.5.0
Django==2.1.2
django-allauth==0.38.0
django-cors-headers==2.4.0
django-rest-auth==0.9.3
django-rest-framework==0.1.0
djangorestframework==3.8.2
idna==2.7
oauthlib==2.1.0
python3-openid==3.1.0
pytz==2018.5
requests==2.20.1
requests-oauthlib==1.0.0
six==1.11.0
urllib3==1.24.1